+++ src/database/index.ts
@@ -1,1 +1,2 @@
 export * as osv from "./osv.js";
+export * as snyk from "./snyk.js";
